requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
azure-core==1.30.2
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-schema==4.16.1
botframework-connector==4.16.1
botframework-streaming==4.16.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
distro==1.9.0
frozenlist==1.4.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
jsonpickle==1.4.2
lxml==5.2.2
msal==1.30.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
openai==1.38.0
pinecone==5.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
yarl==1.9.4
